Creacion de estructura de clases
